// types.ts
export enum Role {
  ADMIN = 'Administrador',
  COORDINATOR = 'Coordinador académico',
  TEACHER = 'Profesor',
  STUDENT = 'Alumno',
  FINANCIAL_MANAGER = 'Gestor Financiero',
}
// ...
export interface AlternativePriceConfig {
    active: boolean;
    monthly: { active: boolean; price: number };
    quarterly: { active: boolean; price: number };
    single: { active: boolean; price: number };
    materialPrice: number;
    enrollmentPrice: number;
}

/**
 * Representa un curso ofrecido por la academia.
 */
export interface Course {
    id: number;
    name: string;
    alternativeName?: string;
    description: string;
    level: string; // Corresponde a CourseLevel.name
    duration: string;
    teacherId: number;
    secondaryTeacherId?: number;
    modality: 'Presencial' | 'Online' | 'Híbrido';
    scheduleIds: number[]; // Array de IDs de WeekSchedule
    classroomId: number;
    minCapacity: number;
    maxCapacity: number;
    status: 'Activo' | 'Archivado' | 'Completado';
    startDate: string; // Cadena de fecha ISO
    endDate: string; // Cadena de fecha ISO
    observations?: string;
    onlineAllowed?: boolean;
    onlineLimit?: number;
    allowWaitingList?: boolean; // Nuevo campo
    paymentRequired?: boolean;
    renewalCourseId?: number;
    receiptGeneration?: string;
    /** Campo calculado: número de clases generadas para este curso. */
    classesCount: number;
    /** Campo calculado: número de recursos asociados a este curso. */
    resourcesCount: number;
    isActive: boolean;
    /** Campo calculado: número de alumnos en lista de espera para este curso. */
    standbyStudents?: number;
    
    alternativePrice?: AlternativePriceConfig;
}
// ...
export interface CourseClass {
  id: string;
  courseId: number;
  date: Date;
  startTime: string; // "HH:MM"
  endTime: string; // "HH:MM"
  teacherId: number;
  isSubstitution: boolean;
  status: 'Hecha' | 'Pendiente' | 'Anulada';
  internalComment?: string;
  publicComment?: string;
  lessonId?: number;
  modality?: 'Presencial' | 'Online';
  classroomId?: number;
}